execution/commitment: consolidate fragmented test files
50 test files in the package, 20 of which held only 1-3 test functions each.
Merge the four clearest clusters, 50 -> 36 files:

  deepfold_test.go     <- 8 files (deepfold_*, deep_storage, storage_fold_navpath)
  witness_test.go      <- 5 files (prune, tracer, exclusion, parity, strict)
  keys_nibbles_test.go <- 2 files merged in place (cache, nibble_addr)
  branch_test.go       <- 2 files (branch_decode, contracthash_prefix)

Pure move: no test function renamed, no test converted to a subtest, no
assertion or helper body changed, so -run filters and CI history still
resolve. 328 Test/Fuzz/Benchmark functions before and after, verified by
diffing the sorted function-name sets. Benchmarks stay in their own files.
