+++ execution/commitment/branch_test.go
@@ -17,9 +17,12 @@
 package commitment
 
 import (
+	"math/rand"
 	"testing"
 
 	"github.com/stretchr/testify/require"
+
+	"github.com/erigontech/erigon/execution/commitment/nibbles"
 )
 
 // TestDecodeBranchInto_RoundTrip asserts that DecodeBranchInto recovers
@@ -85,4 +88,51 @@ func TestDecodeBranchInto_TruncatedInput(t *testing.T) {
 	// Bitmap claims one cell but data missing — should fail.
 	_, err = DecodeBranchInto([]byte{0x00, 0x01}, false, &cells)
 	require.Error(t, err, "bitmap with set bit but no cell data should error")
+}
+
+// reference is the pre-optimization implementation kept as the oracle: it
+// materializes the full hex expansion via CompactToHex, then repacks the first
+// 32 bytes. The zero-alloc ContractHashFromPrefix must agree with it exactly.
+func contractHashFromPrefixReference(prefix []byte) (hash [32]byte, ok bool) {
+	if len(prefix) < 33 {
+		return hash, false
+	}
+	nib := nibbles.CompactToHex(prefix)
+	if len(nib) < 64 {
+		return hash, false
+	}
+	for i := range 32 {
+		hash[i] = nib[2*i]<<4 | nib[2*i+1]
+	}
+	return hash, true
+}
+
+func TestContractHashFromPrefix_MatchesReference(t *testing.T) {
+	rng := rand.New(rand.NewSource(1))
+	for range 5000 {
+		l := 30 + rng.Intn(40) // spans below and above the 33-byte minimum
+		prefix := make([]byte, l)
+		rng.Read(prefix)
+		wantHash, wantOK := contractHashFromPrefixReference(prefix)
+		gotHash, gotOK := ContractHashFromPrefix(prefix)
+		require.Equalf(t, wantOK, gotOK, "ok mismatch len=%d prefix0=%#x", l, prefixByte0(prefix))
+		require.Equalf(t, wantHash, gotHash, "hash mismatch len=%d prefix0=%#x", l, prefixByte0(prefix))
+	}
+}
+
+func prefixByte0(p []byte) byte {
+	if len(p) == 0 {
+		return 0
+	}
+	return p[0]
+}
+
+func TestContractHashFromPrefix_ZeroAlloc(t *testing.T) {
+	prefix := make([]byte, 40)
+	prefix[0] = 0x10 // odd flag set, to exercise the shifting branch
+	allocs := testing.AllocsPerRun(1000, func() { _, _ = ContractHashFromPrefix(prefix) })
+	require.Zero(t, allocs, "ContractHashFromPrefix must not allocate")
+	prefix[0] = 0x00 // even branch
+	allocs = testing.AllocsPerRun(1000, func() { _, _ = ContractHashFromPrefix(prefix) })
+	require.Zero(t, allocs, "ContractHashFromPrefix (even) must not allocate")
 }
